requirements.txt
asgiref==3.8.1
bleach==6.2.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cloudinary==1.43.0
cryptography==44.0.2
debugpy==1.8.13
dj-database-url==2.3.0
Django==5.0
django-allauth==65.7.0
django-cloudinary-storage==0.3.0
django-environ==0.12.0
gunicorn==23.0.0
idna==3.10
packaging==24.2
pillow==11.1.0
psycopg-binary==3.2.6
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.1
tzdata==2025.2
Unidecode==1.3.8
urllib3==2.3.0
webencodings==0.5.1
whitenoise==6.9.0
